refactoring modules * changed module name from luxDataFrame to comply with pandas standard pd.core.frame * changed module name from compiler to processor to encapsulate compiler, parser, and validator
