+++ doc/source/reference/gen/lux.processor.Compiler.Compiler.rst
@@ -1,5 +1,5 @@
 lux.processor.Compiler.Compiler
-==============================
+===============================
 
 .. currentmodule:: lux.processor.Compiler
