+++ doc/source/reference/gen/lux.vis.Vis.Vis.rst
@@ -33,4 +33,10 @@ lux.vis.Vis.Vis
 
    
    
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~Vis.get_data
+
