+++ lux/action/enhance.py
@@ -1,6 +1,6 @@
 import lux
 from lux.interestingness.interestingness import interestingness
-from lux.compiler.Compiler import Compiler
+from lux.processor.Compiler import Compiler
 from lux.utils import utils
 
 #for benchmarking
@@ -11,7 +11,7 @@ def enhance(ldf):
 
 	Parameters
 	----------
-	ldf : lux.luxDataFrame.LuxDataFrame
+	ldf : lux.core.frame
 		LuxDataFrame with underspecified intent.
 
 	Returns
